+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # 0.1.1 (unreleased)
 * bugfix telemetry: record pid also for `request_processing_duration`
+* support abandoning runs on the server (already part of server protocol since 0.1.0)
 
 
 ## 0.1.0 (released on 2024-09-18)