+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # 0.1.1 (unreleased)
 * bugfix telemetry: record pid also for `request_processing_duration`
 * support abandoning runs on the server (already part of server protocol since 0.1.0)
+* support abandoning old runs in `client.py`
+* support `Agent` subclassing along with agent functions in `client.py`
 
 
 ## 0.1.0 (released on 2024-09-18)