+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-# 0.1.1 (unreleased)
+## 0.1.1 (released on 2024-10-02)
 * bugfix telemetry: record pid also for `request_processing_duration`
 * support abandoning runs on the server (already part of server protocol since 0.1.0)
 * support abandoning old runs in `client.py`